+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/Live/LiveInfo/Users/Editors/Editors.tsx
@@ -1,0 +1,71 @@
+import Margin from '@codesandbox/common/lib/components/spacing/Margin';
+import Tooltip from '@codesandbox/common/lib/components/Tooltip';
+import React from 'react';
+import FollowIcon from 'react-icons/lib/io/eye';
+import UnFollowIcon from 'react-icons/lib/io/eye-disabled';
+import RemoveIcon from 'react-icons/lib/md/remove';
+
+import { useSignals, useStore } from 'app/store';
+
+import { SubTitle } from '../../elements';
+
+import { IconContainer, Users } from '../elements';
+import { User } from '../User';
+
+export const Editors = ({ editors }) => {
+  const {
+    live: { onFollow, onRemoveEditorClicked },
+  } = useSignals();
+  const {
+    live: {
+      followingUserId,
+      isOwner,
+      liveUserId,
+      roomInfo: { mode },
+    },
+  } = useStore();
+  const getSideView = userId => (
+    <>
+      {userId !== liveUserId && (
+        <IconContainer>
+          {followingUserId === userId ? (
+            <Tooltip content="Stop following">
+              <UnFollowIcon onClick={() => onFollow({ liveUserId: null })} />
+            </Tooltip>
+          ) : (
+            <Tooltip content="Follow along">
+              <FollowIcon onClick={() => onFollow({ liveUserId: userId })} />
+            </Tooltip>
+          )}
+        </IconContainer>
+      )}
+
+      {isOwner && mode === 'classroom' && (
+        <IconContainer style={{ marginLeft: '0.25rem' }}>
+          <Tooltip content={'Make spectator'}>
+            <RemoveIcon
+              onClick={() => onRemoveEditorClicked({ liveUserId: userId })}
+            />
+          </Tooltip>
+        </IconContainer>
+      )}
+    </>
+  );
+
+  return (
+    <Margin top={1}>
+      <SubTitle>Editors</SubTitle>
+
+      <Users>
+        {editors.map(user => (
+          <User
+            key={user.id}
+            sideView={getSideView(user.id)}
+            type="Editor"
+            user={user}
+          />
+        ))}
+      </Users>
+    </Margin>
+  );
+};
