+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/Live/LiveInfo/Users/Editors/index.ts
@@ -1,0 +1,1 @@
+export { Editors } from './Editors';
