+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/Live/LiveInfo/Users/OtherUsers/index.ts
@@ -1,0 +1,1 @@
+export { OtherUsers } from './OtherUsers';
