+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/Live/LiveInfo/Users/Owners/index.ts
@@ -1,0 +1,1 @@
+export { Owners } from './Owners';
