+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/Live/LiveInfo/Users/index.ts
@@ -1,0 +1,1 @@
+export { Users } from './Users';
